contracts: move CLI output DTOs to IPC crate ( W3 )
